Move references to new package
