+++ src/org/ballerinalang/plugins/idea/psi/references/CallableUnitNameReference.java
@@ -14,20 +14,23 @@
  *  limitations under the License.
  */
 
-package org.ballerinalang.plugins.idea.psi;
+package org.ballerinalang.plugins.idea.psi.references;
 
 import com.intellij.psi.PsiElement;
+import org.ballerinalang.plugins.idea.psi.ConnectorNode;
+import org.ballerinalang.plugins.idea.psi.FunctionNode;
+import org.ballerinalang.plugins.idea.psi.IdentifierPSINode;
 import org.jetbrains.annotations.NotNull;
 
-public class VariableReference extends BallerinaElementReference {
+public class CallableUnitNameReference extends BallerinaElementReference {
 
-    public VariableReference(@NotNull IdentifierPSINode element) {
+    public CallableUnitNameReference(@NotNull IdentifierPSINode element) {
         super(element);
     }
 
     @Override
     public boolean isDefinitionNode(PsiElement def) {
-        return def instanceof VariableDefinitionNode || def instanceof ParameterNode;
+        return def instanceof FunctionNode || def instanceof ConnectorNode;
     }
 
     @NotNull
